Pin PyPi's certificates for pip operations, minor pip_update.sh improvements
Since pip packages often lack signatures, this is one of a number of defenses
soon to be implemented with the intention of hardening the securedrop_app_code
build process. It is important to note that a hardened build process beings at
running pip_update.sh, where defense like cert pinning can help protect against
package rollback and freeze attacks in the pip-compile process through a MitM.
It is of course also needed to harden main package build process carried out by
the build-securedrop-app-code-deb-pkg role (whose name could maybe be
shortened). Not to let a good cert go to waste, however, I've added it to the
app-test and development roles as well.

It should be noted PyPi has used the same cert at least since Dec 2013 (see
github.com/pypa/pip/blob/1.5.6/pip/_vendor/requests/cacert.pem#L1896-L1925), so
I don’t think we need to worry about changing this frequently if we want to put
it in version control, which I have. Since it’s included in the pip package and
the pip is signed, I was able to verify the PGP signature and then I added it to
version control with a signed commit.

I've also added a little bit more functionality to the pip_update.sh script,
just to help things go a little bit smoother and more quickly. The reason for
adding the cleanup function is not because cert pinning has ever caused an
error (and if it did we'd want to know), rather, it is because sometimes other
errors such as ones caused by networking problems require one to manually
cleanup after the script in order to try to run it cleanly again.