Encryption plaintexts: encode UTF-8 and/or make stream
Since GPG operates on raw bytes, plaintext encryption input must be converted to
a stream/file-like type before it can be processed. `python-gnupg` handles this
with a function `gnupg._util._make_binary_stream` that we have taken and
modified as `crypto_util._make_binary_stream` to be more specific to our wants.
Namely, when something is not one of `python-gnupg`'s recognized
`gnupg._util._STREAMLIKE_TYPES`, we encode it as a UTF-8 `str` (unless it's
already a `str`), then always convert it to a `_io.BytesIO` stream.

This is opposed to `python-gnupg`'s method of choosing the encoding based on
locale settings, which we have observed to be problematic in conjunction with
how Apache creates Apache-controlled environments for WSGI applications, such as
SD (see https://github.com/freedomofpress/securedrop/issues/1343 and
https://github.com/freedomofpress/securedrop/issues/1360--duplicate issues).
Explicitly controlling the encoding and type of input we send to
`gnupg.GPG.encrypt()` will fix these two issues, and should prevent new ones
related to encoding of plaintexts.

`python-gnupg` also calls `gnupg._util._make_binary_stream()` on ciphertext
inputs to `gnupg.GPG.decrypt()` and batch file inputs `gnupg.GPG.gen_key()`
(without even testing that the input's type belongs to `_STREAMLIKE_TYPES` with
`gnupg._util._is_stream`). Ciphertexts are decrypted when a source logs in,
where the ciphertext is the result of calling the `read()` method on an open
`file`, which will produce a `str`. In our tests, we also pass `str`s as
ciphertexts. If we were to convert these ciphertext `str`s to `_io.BytesIO`
streams before passing to `python-gnupg` for decryption,
`gnupg._util._make_binary_stream` would throw an exception because it tries to
call the `encode()` on all non-`str` types. So, due to this non-standardized
handling between the encrypt and decrypt methods, we simply let `python-gnupg`
convert our `str`s (which it won't try to `encode`, as it shouldn't) to
`_io.BytesIO` streams.  `gnupg._util.gen_key_input()` always returns a `str`.
So, for the same reasons as with ciphertexts, we do not do any additional
processing on the output of this method (a batch file input for
`gnupg._util.gen_key()`).

It may be wise to try to make a pull request upstream to standardize the
behaviour of `python-gnupg`. It would be ideal to only pass `_io.BytesIO` stream
objects to `python-gnupg` in order to be most certain we won't run into further
errors with `python-gnupg` processing our plaintext input objects, and instead
they will be passed (more) directly to GPG.

(Also, this commit correctly identifies the object of decryption as the
ciphertext.)