Add unit test to ensure UTF-8 encoding despite env locale settings
This is a follow up to 56f23087303d3fd207c2444e47c764604ae619dc to prevent
regressions. It fixes the unicode test to emulate the current production
environment, where gnupg-python initializes our `GPG` objects with
ansi_x3.4_1968 encoding due to the fact Apache launches WSGI applications in an
Apache-controlled environment (see
https://github.com/freedomofpress/securedrop/issues/1360#issuecomment-239593607).
The fact this encoding attribute was not properly set in our testing environment
which doesn't use Apache led to the re-introduction of the bug that
56f23087303d3fd207c2444e47c764604ae619dc fixed.

The bug originated in 37ba11c7b94f1a547aa5e8ed342e3da85a663362 and was the
result of the fact at that time I didn't realize Apache launched WSGI in a
custom environment (with LANG=C by default). If this weren't the case then
python-gnupg would have no problems with encoding. As noted in
37ba11c7b94f1a547aa5e8ed342e3da85a663362, however, it makes the most sense to
explicitly handle this ourselves as that is the most robust way to guard against
future regressions.

As noted in the comment added, we couldn't do this exactly the way we'd like due
to the current setup of our test suite. However, through unsavory
monkey-patching we've achieved the desired effect.

Also, I changed the UTF-8 test string because Anglo-Saxon rune poems look
awesome.